requirements.txt
python-telegram-bot==13.15
apscheduler==3.10.4
pytz==2023.3